comments modified

+++ week3/week3/src/week3/Factorial.java
@@ -15,7 +15,7 @@ public class Factorial {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         int number = input.nextInt();
-        double  result=1; // chose double data type because there very larg numbers in factorial
+        long  result=1; 
         for (int i = number; i >1 ; i--) {
             result*=i;
         }
